+++ Content.Shared/SimpleStation14/Magic/Components/ClothingGrantComponentComponent.cs
@@ -3,8 +3,10 @@ namespace Content.Shared.SimpleStation14.Magic
     [RegisterComponent]
     public sealed class ClothingGrantComponentComponent : Component
     {
-        [DataField("component", required: true), ViewVariables(VVAccess.ReadWrite)]
-        public string Component = "";
+        [DataField("component"), ViewVariables(VVAccess.ReadWrite)]
+        public string? Component = null;
+        [DataField("tag"), ViewVariables(VVAccess.ReadWrite)]
+        public string? Tag = null;
 
         [ViewVariables(VVAccess.ReadWrite)]
         public bool IsActive = false;
