+++ Resources/Prototypes/SimpleStation14/tags.yml
@@ -1,0 +1,2 @@
+- type: Tag
+  id: PlasmamanFlammable
